Default rows for HYPREMatrix(::SparseMatrixCS(C|R)) and HYPREVector(::Vector)
This patch adds defaults for ilower and iupper in constructors when i)
using SparseMatrixCS(C|R) and Vector and ii) when no communicator is
passed. In this case we simply assume the users wants a single-processor
solve, and default to passing 1 and size(A, 1) as the default values.

In the future this could possibly use the HYPRE sequential compiler
flag to completely avoid MPI.